use ES6 export, export  all directives as default export, add default export tests
Updated all files (except for src/directiveLocation) to use ES6 export syntax.

Changed the default export for the library to be an array of all directives to streamline usage in situations where a user wants to use all of the directives provided by the library.

Added a new test to ensure the default export is an array containing only instances of GraphQLDirective.

Updated README examples with new usage.